Move handler descriptions to separate files
